feat: Add Digital Asset Links file
This commit adds an `assetlinks.json` file to the `.well-known` directory.

This file is used for Digital Asset Links, which enable Android apps to claim ownership of web domains.
The file declares that the `cz.jaro.dpmcb` and `cz.jaro.dpmcb.debug` Android packages are authorized to handle all URLs associated with the domain.