feat: Remove direct usage of Firebase Crashlytics
This commit removes direct usage of Firebase Crashlytics.
It introduces `recordException` and `log` functions in `crashlytics.android.kt` to centralize error reporting.
All previous calls to `Firebase.crashlytics.recordException` and `Firebase.crashlytics.log` are replaced with calls to the new `recordException` and `log` functions.
This change makes crash reporting easier to manage and update in the future.