feat: Add URL encoding function
This commit introduces a new function `encodeURL` for encoding strings to be used in URLs.
It also moves the URL encoding logic from `MainViewModel` into this new function, which now resides in `otherUtils`.
The implementation of this function is platform-specific, using `URLEncoder` on Android and `encodeURLPath` on JS.
This change centralizes URL encoding and improves code maintainability.